An example change for tests.
Here I've pulled the latest main branch, then branched off into this
branch `fixup_...`; I've made a change and documented it, added it
to the staging, and then committed it to the branch. I'll then push it
to upstream and potentially open a pull request so you all can comment
on it.